blockd: also send ubus notification on mount hotplug
Also sending a ubus notification on mount hotplug provides a useful
shortcut for procd service triggers.
As the /etc/hotplug.d/mount API for now doesn't have any users and also
makes it hard to aggregate hotplug calls, we should consider removing it
in favor of only using triggers on the block notifications in future.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>