blockd: move to its own POSIX process group
Not to be confused with cgroups, there are also POSIX process groups.
They do matter when it comes to autofs, as all requests coming from the
process group of the automounter itself will be ignored.
Hence, if blockd runs in the same process group as init and all
services, requests from services will be ignored.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>