libfstools: add uImage.FIT fitblk driver
The upcoming 'fitblk' block driver exposing the filesystem sub-image(s)
of a uImage.FIT storage on a block device as /dev/fit%u.
In case of the block device being a fixed-sized partition, the remaining
space (for rootfs_data) is mapped to /dev/fitrw.
Recognize both devices and assign the highest priority to the new driver.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>