libfstools:support to specify filesystem when formatting overlay
The old strategy uses F2FS when the block device size > F2FS_MINSIZE, otherwise it uses EXT4.
This strategy is not very flexible, because users usually don’t choose the filesystem just based on the size of the device,
so we need to add a macro to let users decide which filesystem to use.

use the "-DCMAKE_BLKDEV_FILESYSTEM_TYPE=EXT4" option to select EXT4 file system

Signed-off-by: zexiong <xiezexiong@foxmail.com>