libfstools: skip JFFS2 padding when BLOCKSIZE was given
When BLOCKSIZE is defined we call padjffs2 with the defined alignment
to pad towards. That results in one of more block of all-1s instead of
the expected 0xdeadc0de and ultimaltely breaks restoring configuration
as fstools won't look for a the GZ signature indicating the config
backup in this case.
Change that and also look for config backup when all-1s are used for
padding.

Fixes: ee54c6b ("libfstools: skip JFFS2 padding on block devices")
Signed-off-by: Daniel Golle <daniel@makrotopia.org>